app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.5
ollama==0.3.3
python-dotenv==1.0.1
pydantic==2.9.2
jinja2==3.1.4
